Create Js
This code adds a PDF file upload field and a button to convert the file to an Excel spreadsheet. When the button is clicked, the PDF file is loaded using the pdf.js library and the PDF pages are extracted as text. The text is then formatted into an array of arrays for the sheet.js library. Once all PDF pages are extracted and formatted into an Excel spreadsheet, the file is saved as a blob in the browser and a download link is created for the user to download the file.